pyproject.toml
[tool.poetry]
name = "bridge"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.0"
azure-storage-blob = "^12.19.0"
msal = "^1.25.0"
flask-session = "^0.5.0"
werkzeug = "^3.0.1"
flask-caching = "^2.1.0"
markupsafe = "^2.1.3"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "bink-pypi"
url = "https://pypi.gb.bink.com/simple"
priority = "explicit"

[tool.black]
line-length = 120

[tool.ruff]
select = ["ALL"]
ignore = ["ANN101", "ANN001", "COM812", "D203", "D213", "D407", "FA102", "G004", "PLR0913", "S311"]
line-length = 250